refactor(BA-6609): drop api re-export shims for the relocated primitives
Point every consumer directly at the data-layer definitions and remove
the backward-compatibility re-exports added in the previous commit:

- Undefined/undefined -> data.common.sentinel
- ManagerStatus -> data.manager_status.types
- BaseFilterAdapter -> data.filter.adapter

api.utils and api.__init__ no longer re-export the moved symbols, and
the now-empty api/rest/adapter.py shim is deleted.
