--- src/ai/backend/manager/api/__init__.py
@@ -1,4 +1,0 @@
-# Re-exported for backward compatibility; defined in the data layer.
-from ai.backend.manager.data.manager_status.types import ManagerStatus
-
-__all__ = ["ManagerStatus"]
