+++ src/ai/backend/manager/api/rest/agent/adapter.py
@@ -14,8 +14,8 @@ from ai.backend.common.dto.manager.agent import (
     OrderDirection,
     SearchAgentsRequest,
 )
-from ai.backend.manager.api.rest.adapter import BaseFilterAdapter
 from ai.backend.manager.data.agent.types import AgentDetailData, AgentStatus
+from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.repositories.agent.query import QueryConditions, QueryOrders
 from ai.backend.manager.repositories.base import BatchQuerier, OffsetPagination
