+++ src/ai/backend/manager/api/rest/auto_scaling_rule/adapter.py
@@ -19,9 +19,9 @@ from ai.backend.common.dto.manager.auto_scaling_rule import (
     UpdateAutoScalingRuleRequest,
 )
 from ai.backend.common.types import AutoScalingMetricSource
-from ai.backend.manager.api.rest.adapter import BaseFilterAdapter
 from ai.backend.manager.data.deployment.scale_modifier import ModelDeploymentAutoScalingRuleModifier
 from ai.backend.manager.data.deployment.types import ModelDeploymentAutoScalingRuleData
+from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.endpoint.conditions import AutoScalingRuleConditions
 from ai.backend.manager.models.endpoint.orders import AutoScalingRuleOrders
