+++ src/ai/backend/manager/api/rest/domain/adapter.py
@@ -16,8 +16,8 @@ from ai.backend.common.dto.manager.domain import (
     UpdateDomainRequest,
 )
 from ai.backend.common.types import ResourceSlot
-from ai.backend.manager.api.rest.adapter import BaseFilterAdapter
 from ai.backend.manager.data.domain.types import DomainData
+from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.domain import DomainRow
 from ai.backend.manager.models.domain.conditions import DomainConditions
