+++ src/ai/backend/manager/api/rest/notification/adapter.py
@@ -30,7 +30,7 @@ from ai.backend.common.dto.manager.notification import (
     WebhookSpecResponse,
 )
 from ai.backend.common.dto.manager.notification.response import EmailSpecResponse
-from ai.backend.manager.api.rest.adapter import BaseFilterAdapter
+from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.data.notification import (
     NotificationChannelData,
     NotificationRuleData,
