+++ src/ai/backend/manager/api/rest/prometheus_query_preset/adapter.py
@@ -19,7 +19,7 @@ from ai.backend.common.dto.manager.prometheus_query_preset import (
     QueryDefinitionOrderField,
     SearchQueryDefinitionsRequest,
 )
-from ai.backend.manager.api.rest.adapter import BaseFilterAdapter
+from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.data.prometheus_query_preset import PrometheusQueryPresetData
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.prometheus_query_preset import PrometheusQueryPresetRow
