+++ src/ai/backend/manager/config/loader/legacy_etcd_loader.py
@@ -12,7 +12,7 @@ from ai.backend.common.configs.loader.types import AbstractConfigLoader
 from ai.backend.common.etcd import AsyncEtcd, Event, GetPrefixValue
 from ai.backend.common.identity import get_instance_id
 from ai.backend.common.types import QueueSentinel, SlotName, SlotTypes, current_resource_slots
-from ai.backend.manager.api import ManagerStatus
+from ai.backend.manager.data.manager_status.types import ManagerStatus
 from ai.backend.manager.defs import INTRINSIC_SLOTS
 from ai.backend.manager.errors.common import ServerMisconfiguredError
