+++ tests/unit/manager/api/test_adapter.py
@@ -13,7 +13,7 @@ import sqlalchemy as sa
 
 from ai.backend.common.data.filter_specs import StringInMatchSpec, StringMatchSpec
 from ai.backend.common.dto.manager.query import StringFilter
-from ai.backend.manager.api.rest.adapter import BaseFilterAdapter
+from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.models.clauses import QueryCondition
 
 if TYPE_CHECKING:
